requirements.txt
alembic==1.16.4
asgiref==3.9.1
asttokens==3.0.0
catboost==1.2.8
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
colorlog==6.9.0
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
diff-match-patch==20241021
Django==5.2.5
django-import-export==4.3.9
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fonttools==4.59.1
graphviz==0.21
greenlet==3.2.4
idna==3.10
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lightgbm==4.6.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
missingno==0.5.2
narwhals==2.1.2
nest-asyncio==1.6.0
numpy==2.2.6
openpyxl==3.1.5
optuna==4.5.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
patsy==1.0.1
pillow==11.3.0
platformdirs==4.3.8
plotly==6.3.0
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.1
requests==2.32.5
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
statsmodels==0.14.5
tablib==3.8.0
threadpoolctl==3.6.0
tomli==2.2.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
xgboost==3.0.4
